broadwell: Clean up XHCI and EHCI ramstage drivers
The EHCI controller is disabled in refcode binary in the typical case,
it is only left enabled when using USBDEBUG.  So the code to disable
the controller can be removed.

The XHCI controller clock gating setup is done in the ramstage
refcode binary so that code can be removed now.  The SMM code for
preparing the controller prior to S3/S5 is reworked but still
made available to be called by the SMI handler.

BUG=chrome-os-partner:28234
TEST=None

Change-Id: I19d1df99d2ee5dbc9c93ca01b2d246432928d77f
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/199181
Reviewed-by: Aaron Durbin <adurbin@chromium.org>