broadwell: Clean up gpio handling code
Only the "LP" variant interface to GPIOs is supported so drop
all the LP_ prefixes and remove the special handling for the
previous GPIO interface model.

Change all occurrences of get_gpiobase() to GPIO_BASE_ADDRESS.

Add a new function to initialize a single GPIO which can be used
by a mainboard to support specific setup and/or power sequence
requirements.

BUG=chrome-os-partner:28234
TEST=None

Change-Id: I8f645f914eb576e00b3c8feb93c8291d763640d0
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/199182
Reviewed-by: Aaron Durbin <adurbin@chromium.org>