broadwell: Add romstage code to configure PCH UART for console
The two UARTs built into the haswell/broadwell PCH can be used for
standard console output if configured properly.  There is a magic bit
in the IOBP settings for the UART device that will put it into "byte
addressing mode" so it is compatible with standard 16550 UART drivers.

When in this mode the linux kernel driver will be unable to talk to
the device so it is indicated as disabled via the ACPI driver.

Note that this by itself is not enough to get working MMIO UART
in coreboot because the current code is heavily tied with the oxpcie
driver.  I have a separate set of patches to disentangle the generic
MMIO UART code from the oxpcie driver but I want to get that series
in upstream first.

BUG=chrome-os-partner:28234
TEST=None

Change-Id: I9e689456aa5017784328d1be33ad072f79db1920
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/199807